% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{cmetEntropicCoords}
\alias{cmetEntropicCoords}
\title{A caracterization of the variables in a data.frame of channel multivariate entropies}
\format{An object of class \code{character} of length 5.}
\usage{
cmetEntropicCoords
}
\description{
A caracterization of the variables in a data.frame of channel multivariate entropies
}
\keyword{datasets}

