% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condentropies.R
\name{condentropies.data.frame}
\alias{condentropies.data.frame}
\title{Entropy decomposition of a data frame (multivariate decomposition)}
\usage{
\method{condentropies}{data.frame}(X, Y, ...)
}
\value{
Another dataframe with the main entropy coordinates of every variable Xi
  in the original conditioned on the datatables Y, which are now the rows of the returned data.frame.
}
\description{
Entropy decomposition of a data frame (multivariate decomposition)
}
