% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condentropies.R
\name{condentropies}
\alias{condentropies}
\title{Conditional entropy decomposition of dataframes, similar to entropies}
\usage{
condentropies(X, Y, ...)
}
\arguments{
\item{data}{The data being provided to the function.}
}
\value{
A dataframe with the entropies of the marginals
}
\description{
Returns several different flavours of entropies depending on the structure 
the data is provided to the function. There are specialized versions for
(contingency) tables, confusion matrices and data frames.
}
\details{
Unless specified by the user explicitly, this function uses base 2 
  logarithms for the entropies.
}
\seealso{
\code{\link[entropy]{entropy}, \link[infotheo]{entropy}}
}

