% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{derivedSplitEntropies}
\alias{derivedSplitEntropies}
\title{A caracterization of the variables in a data.frame of derived non-split entropies.}
\format{\preformatted{ chr [1:6] "DeltaHx" "DeltaHy" "MIxyX" "MIxyY" "VIx" "VIy"
}}
\usage{
derivedSplitEntropies
}
\description{
A caracterization of the variables in a data.frame of derived non-split entropies.
}
\keyword{datasets}

