% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropicCoords.R
\name{entropicCoords}
\alias{entropicCoords}
\title{Transform entropies into entropic coordinates}
\usage{
entropicCoords(df, split = FALSE, purge = FALSE, norm = TRUE)
}
\arguments{
\item{df}{A data frame containing values for the entropies related to a joint
distribution or sets of them.}

\item{split=FALSe}{A flag to select obtaining split coordinates.}

\item{purge=FALSE}{A flag to purge the entropies in \code{df}. If \code{purge=FALSE}
the new entropies are added to the input entropies, otherwise they are purged.}

\item{norm=TRUE}{A flag to request normalized coordinates. If norm=TRUE, the 
coordinates, either split or not, are normalized and ready to plot in the 
entropy triangle.}
}
\value{
A data frame with the entropy points for
}
\description{
Generate the entropic coordinates from the component entropies of a joint 
  count distribution or a contingency matrix.
- with split=TRUE, generate independen coordinates for X and Y distributions.
}
\examples{
coords <- entropicCoords(entropies(UCBAdmissions))
# Note there is almost no information being transmitted
unnormCoords <- entropicCoords(entropies(UCBAdmissions), norm=FALSE)
all(apply(unnormCoords[8:10], 1, sum) == unnormCoords$Uxy)
entropicCoords(entropies(UCBAdmissions), purge=TRUE)
# To obtain the split entropies (In this case the normalization makes no sense)
entropicCoords(entropies(UCBAdmissions), purge=TRUE, split=TRUE)
}
\seealso{
\code{entropies}
}

