% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropies.R
\name{entropies.confusionMatrix}
\alias{entropies.confusionMatrix}
\title{Entropy decomposition of a confusion table}
\usage{
\method{entropies}{confusionMatrix}(ct, ...)
}
\description{
Entropy decomposition of a confusion table
}

