% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropies.R
\name{entropies}
\alias{entropies}
\title{Entropy decomposition of a contingency matrix}
\usage{
entropies(data, ...)
}
\arguments{
\item{N}{An n-contingency matrix where n > 2}

\item{unit}{The logarithm to be used in working out the entropies as per
\code{entropy}. Defaults to "log2".}
}
\value{
A dataframe with the entropies of the marginals
}
\description{
Given a contingency matrix, provide one row of entropy coordinates.
NOTE: the reference variable has to index the ROWS of the table, while the predicted
variable indexes the columns, unlike, e.g. \code{\link[caret]{contingencyTable}}
}
\details{
Unless specified by the user explicitly, this function uses base 2
logarithms for the entropies.
}
\seealso{
\code{\link[entropy]{entropy}}
}

