% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entropies.R
\name{entropies.table}
\alias{entropies.table}
\title{Entropy decomposition of a contingency matrix}
\usage{
\method{entropies}{table}(Nxy, ...)
}
\description{
Given a contingency matrix, provide one row of entropy coordinates.
}

