% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{A function to evaluate confusion matrices and contingency tables.}
\usage{
evaluate(data, ...)
}
\arguments{
\item{data}{Either a \code{\link[base]{table}} working as an n-way contingency 
table or a \code{\link[caret]{confusionMatrix}}}
}
\value{
At least, the concatenation of the entropies (whether split or not), the \code{EMA}
 and \code{NIT} rate as explained in \cite{val:pel:14a}.
}
\description{
Provides also the EMA and NIT over that provided by perplexities.
}
