% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.table}
\alias{evaluate.table}
\title{A primitive to evaluate a contingency table}
\usage{
\method{evaluate}{table}(cm, ...)
}
\arguments{
\item{cm}{A contingency table as per \code{\link[base]{table}}}

\item{split=FALSE}{Whether to evaluat the split entropies or not.}
}
\description{
The criteria evaluated at present are:
\itemize{
\item  entropic quantities, as issued from \code{entropies} and \code{entropicCoords}.
\item perplexities, as issued from \code{\link{perplexities}}
}
}
\examples{
evaluate.table(UCBAdmissions) # from package datasets
}

