% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherCoords.R
\name{gatherCoords}
\alias{gatherCoords}
\title{A function to massage the split coordinates back to non-split with an indication of origin}
\usage{
gatherCoords(data, limit = TRUE)
}
\arguments{
\item{data}{The data in splitEntropies format.}
}
\description{
A function to massage the split coordinates back to non-split with an indication of origin
}
