% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggentropytern.R
\name{ggentropytern}
\alias{ggentropytern}
\title{Entropic ternary plot constructor}
\usage{
ggentropytern(data, ...)
}
\arguments{
\item{experiments}{A dataframe with the data to be plotted, typically obtained with \code{entropicCoordinates}.
To plot the *split entropy triangle* this dataframe must have  at least the following variables \code{c("DeltaHx", "DeltaHx","MIxyX","MIxyY", "VIx", "VIy")}.
To plot the *entropy triangle* this dataframe must have  at least the following variables: 
\code{c("DeltaHxy", "MIxy2", "VIxy")}.}
}
\value{
A plot object in the ggplot2 class of plots ready to be plotted.
}
\description{
This function creates a basic ternary plot on entropic coordinates whether 
split or not.

 @details The axes of the plot are changed according to whether the three
aggregate variables "DeltaHxy", "MIxy2" and "VIxy" are defined (not split) or
the split variables "DeltaHx", "DeltaHx","MIxyX", "MIxyY", "VIx", "VIy" are defined in
the dataframe passed as parameters. 
  Note that no geometry is imposed on the plot.
}
\details{
Only the plot and aesthetics are generated, the geometry is not decided.
}
\examples{
data(UCBAdmissions)
experiments <- entropicCoords(entropies(UCBAdmissions)) # Non-split data
eT <- ggentropytern(experiments) + geom_point(aes(VIxy,MIxy2,DeltaHxy, colour=Dept, shape=Dept))
eT + ggtitle("UCB admissions by department. X=admittance status, Y=gender") # non-split entropy triangle

}
\seealso{
\code{\link{ggtern}}, \code{\link{entropies}}, \code{\link{entropicCoordinates}}
}

