% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmetern.R
\name{ggmetern}
\alias{ggmetern}
\title{Grammar-of-Graphics Multivariate entropic ternary plot constructor}
\usage{
ggmetern(ed, fancy = TRUE, ...)
}
\arguments{
\item{ed}{A dataframe with the entropy data to be plotted, typically obtained with \code{\link[infotheo]{entropy}}.
To plot the *split entropy triangle* this dataframe must have  at least the following variables \code{c("DeltaHx", "DeltaHx","MIxyX","MIxyY", "VIx", "VIy")}.
To plot the *entropy triangle* this dataframe must have  either of the two configurations.}

\item{fancy}{A switch whether to use a fancier (coloured) or terser (B&W) \code{\link[ggplot2]{theme}} in the plot.
Defaults to fancy=TRUE, for exploratory analysis. fancy=FALSE is better for printed matter.}
}
\value{
A plot object in the ggplot2 class of plots ready to be plotted.
}
\description{
This function creates a ternary plot on entropic coordinates whether 
split or not for multivariate sets of random variables.
}
\details{
The axes of the plot are changed according to whether the three
aggregate variables "DeltaH_Px", "M_Px" and "VI_Px" are defined (not split) or
the split variables "DeltaH_Pxi", "DeltaH_Pxi","M_Pxi" in the dataframe passed as parameter. 
  Note that no geometry is imposed on the plot.

Only the plot and aesthetics are generated, the geometry is not decided. A further 
column, including the name of the original variables may be used to annotate the plot.
}
\seealso{
\pkg{ggtern}, \pkg{entropies}, \pkg{entropicCoordinates}
}
