% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasCmetEntropicCoords}
\alias{hasCmetEntropicCoords}
\title{A function to detect if the the channel multivariate entropic coordinates are present}
\usage{
hasCmetEntropicCoords(df)
}
\description{
A function to detect if the the channel multivariate entropic coordinates are present
}
