% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{hasDerivedNonSplitEntropies}
\alias{hasDerivedNonSplitEntropies}
\title{A function to detect if the derived, non-split entropies are present.}
\usage{
hasDerivedNonSplitEntropies(df)
}
\description{
A function to detect if the derived, non-split entropies are present.
}

