% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasDerivedSplitEntropies}
\alias{hasDerivedSplitEntropies}
\title{A function to detect if the derived split entropies are present.}
\usage{
hasDerivedSplitEntropies(df)
}
\description{
A function to detect if the derived split entropies are present.
}

