% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasMultiAggregateEntropies}
\alias{hasMultiAggregateEntropies}
\title{a function to detect if the multivariate simple entropies are present: this enables 
working out the multivariate simple entropic coordinates}
\usage{
hasMultiAggregateEntropies(df)
}
\description{
a function to detect if the multivariate simple entropies are present: this enables 
working out the multivariate simple entropic coordinates
}
