% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasMultiEntropicCoords}
\alias{hasMultiEntropicCoords}
\title{A function to detect if the multivariate entropic coordinates are present}
\usage{
hasMultiEntropicCoords(df)
}
\description{
A function to detect if the multivariate entropic coordinates are present
}

