% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasMultiSplitEntropicCoords}
\alias{hasMultiSplitEntropicCoords}
\title{A function to detect if the multivariate split entropic coordinates are present}
\usage{
hasMultiSplitEntropicCoords(df)
}
\description{
A function to detect if the multivariate split entropic coordinates are present
}

