% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasSplitSmetCoords}
\alias{hasSplitSmetCoords}
\title{a function to detect if the multivariate split entropies are present: this enables 
working out the multivariate split entropic coordinates}
\usage{
hasSplitSmetCoords(df)
}
\description{
a function to detect if the multivariate split entropies are present: this enables 
working out the multivariate split entropic coordinates
}

