% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jentropies.R
\name{jentropies.data.frame}
\alias{jentropies.data.frame}
\title{Mutivariate Joint Entropy decomposition of a data frame}
\usage{
\method{jentropies}{data.frame}(X, Y, ...)
}
\value{
Another dataframe with the main entropy coordinates of every variable Xi
  in the original conditioned on the datatables Y, which are now the rows of the returned data.frame.
}
\description{
Mutivariate Joint Entropy decomposition of a data frame
}
