% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jentropies.R
\name{jentropies.table}
\alias{jentropies.table}
\title{Entropy decomposition of a possibly higher-order contingency matrix}
\usage{
\method{jentropies}{table}(Nxy, ...)
}
\arguments{
\item{Nxy}{An n-contingency matrix where n > 2}

\item{unit}{The logarithm to be used in working out the sentropies as per 
\code{\link[entropy]{entropy}}. Defaults to "log2".}
}
\description{
Given a contingency matrix, provide one row of entropy coordinates. 
NOTE: the reference variable has to index the ROWS of the table, while the predicted
variable indexes the columns, unlike, e.g. \code{\link[caret]{confusionMatrix}}. 
That is the entropies are obtained for the first two margins. If other margins are 
needed you will need to reorder them.
}
