% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexities.R
\name{perplexities.confusionMatrix}
\alias{perplexities.confusionMatrix}
\title{A function to obtain perplexities on a \code{\link[caret]{confusionMatrix}}}
\usage{
\method{perplexities}{confusionMatrix}(data, ...)
}
\arguments{
\item{data}{A confusion matrix from an evaluation as in \code{\link[caret]{confusionMatrix}}}
}
\description{
Obtains the perplexities in a confusion matrix by calculating, first, 
the entropies, and then the entropic coordinates for it.
}
