% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexities.R
\name{perplexities.data.frame}
\alias{perplexities.data.frame}
\title{A function to work out the perplexities of a data frame with entropic information in it.}
\usage{
\method{perplexities}{data.frame}(data, ...)
}
\arguments{
\item{A}{dataframe with the entropic variables needed to work out the perplexities, 
namely "Ux", "Uy", "Hx", "Hy" and "Hxy".}
}
\value{
A dataframe with the perplexities associated to the data
}
\description{
A function to work out the perplexities of a data frame with entropic information in it.
}
