% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexities.R
\name{perplexities}
\alias{perplexities}
\title{Obtaining the perplexities (inverse entropies) of a joint distribution.}
\usage{
perplexities(data, ...)
}
\arguments{
\item{data}{The data to evaluate for perplexity.}
}
\value{
A dataframe with the perplexities associated to the data
}
\description{
There are primitives to obtain this for \code{data.frame} and
\code{\link[caret]{confusionMatrix}}.
}
