% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perplexities.R
\name{perplexities.table}
\alias{perplexities.table}
\title{A function to obtain the perplexities of 2- and 3-way tables}
\usage{
\method{perplexities}{table}(data, ...)
}
\arguments{
\item{data}{A 2- or 3-way table.}
}
\description{
If data is a 2-way table, a single line of perplexities is obtained, 
but if data is a 3-way table, then we obtain as many rows as instances in the third
dimension and an extra factor variable on that dimension.
}
\examples{
perplexities(UCBAdmissions)
}
