% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentropies.R
\name{sentropies.confusionMatrix}
\alias{sentropies.confusionMatrix}
\title{Entropy decomposition of a confusion table}
\usage{
\method{sentropies}{confusionMatrix}(ct, ...)
}
\description{
Entropy decomposition of a confusion table
}

