% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentropies.R
\name{sentropies.data.frame}
\alias{sentropies.data.frame}
\title{Multivariate source entropy decomposition of a data frame}
\usage{
\method{sentropies}{data.frame}(df, ...)
}
\value{
Another dataframe with the main entropy coordinates of every variable
  in the original, which are now the rows of the returned data.frame. If the columns have no
  names, artificial ones are returned based in pre prefix "x" and their column number.
}
\description{
Multivariate source entropy decomposition of a data frame
}

