% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentropies.R
\name{sentropies}
\alias{sentropies}
\title{Multivariate Source Entropy decomposition of a dataframe.}
\usage{
sentropies(data, ...)
}
\arguments{
\item{data}{The data being provided to the function.}

\item{unit}{The logarithm to be used in working out the sentropies as per 
\code{entropy}. Defaults to "log2".}
}
\value{
A dataframe with the sentropies of the marginals. If type="dual" then \code{hasAggregateSmetCoords(sentropies(data, type="dual"))}, 
otherwise, \code{hasSplitSmetCoords(sentropies(data))}.
}
\description{
Returns several different flavours of sentropies depending on the structure 
the data is provided to the function. There are specialized versions for
(contingency) tables, confusion matrices and data frames.
}
\details{
Unless specified by the user explicitly, this function uses base 2 
  logarithms for the sentropies.
}
\seealso{
\code{\link[entropy]{entropy}, \link[infotheo]{entropy}}
}
