% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentropies.R
\name{sentropies.table}
\alias{sentropies.table}
\title{Entropy decomposition of a contingency matrix}
\usage{
\method{sentropies}{table}(data, type = "total", ...)
}
\arguments{
\item{Nxy}{An n-contingency matrix where n > 2}

\item{unit}{The logarithm to be used in working out the sentropies as per 
\code{entropy}. Defaults to "log2".}
}
\description{
Given a contingency matrix, provide one row of entropy coordinates. 
NOTE: the reference variable has to index the ROWS of the table, while the predicted
variable indexes the columns, unlike, e.g. \code{\link[caret]{confusionMatrix}}
}
