% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{simpleEntropies}
\alias{simpleEntropies}
\title{A characterization of the variables in a data.frame needed to define the perplexities.}
\format{An object of class \code{character} of length 5.}
\usage{
simpleEntropies
}
\description{
A characterization of the variables in a data.frame needed to define the perplexities.
}
\keyword{datasets}

