% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{derivedNonSplitEntropies}
\alias{derivedNonSplitEntropies}
\title{A caracterization of the variables in a data.frame of derived non-split entropies.}
\format{\preformatted{ chr [1:4] "Uxy" "DeltaHxy" "MIxy2" "VIxy"
}}
\usage{
derivedNonSplitEntropies
}
\description{
A caracterization of the variables in a data.frame of derived non-split entropies.
}
\keyword{datasets}

