% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{derivedSplitEntropies}
\alias{derivedSplitEntropies}
\title{A caracterization of the variables in a data.frame of derived non-split entropies.}
\format{An object of class \code{character} of length 6.}
\usage{
derivedSplitEntropies
}
\description{
A caracterization of the variables in a data.frame of derived non-split entropies.
}
\keyword{datasets}

