% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmetern.R
\name{ggmetern}
\alias{ggmetern}
\title{Grammar-of-Graphics Multivariate entropic ternary plot constructor}
\usage{
ggmetern(data, ...)
}
\arguments{
\item{experiments}{A dataframe with the data to be plotted, typically obtained with \code{entropicCordinates}.
To plot the *split entropy triangle* this dataframe must have  at least the following variables \code{c("DeltaHx", "DeltaHx","MIxyX","MIxyY", "VIx", "VIy")}.
To plot the *entropy triangle* this dataframe must have  either of the two configurations.}
}
\value{
A plot object in the ggplot2 class of plots ready to be plotted.
}
\description{
This function creates a ternary plot on entropic coordinates whether 
split or not for multivariate sets of random variables.

 @details The axes of the plot are changed according to whether the three
aggregate variables "DeltaH_Px", "M_Px" and "VI_Px" are defined (not split) or
the split variables "DeltaH_Pxi", "DeltaH_Pxi","M_Pxi" in the dataframe passed as parameter. 
  Note that no geometry is imposed on the plot.
}
\details{
Only the plot and aesthetics are generated, the geometry is not decided. A further 
column, including the name of the original variables may be used to annotate the plot.
}
\examples{
data(UCBAdmissions)
experiments <- entropicCoords(entropies(UCBAdmissions)) # Non-split data
eT <- ggentropytern(experiments) + geom_point(aes(VIxy,MIxy2,DeltaHxy, colour=Dept, shape=Dept))
eT + ggtitle("UCB admissions by department. X=admittance status, Y=gender") # non-split entropy triangle

}
\seealso{
\code{\link{ggtern}}, \code{\link{entropies}}, \code{\link{entropicCoordinates}}
}

