% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{hasSimpleEntropies}
\alias{hasSimpleEntropies}
\title{a function to detect if the simple entropies are present.}
\usage{
hasSimpleEntropies(df)
}
\description{
a function to detect if the simple entropies are present.
}

