% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{hasSplitSmetCoords}
\alias{hasSplitSmetCoords}
\title{Functions to detect SMET coordinates
a function to detect if the source multivariate split entropies are present: this enables 
working out the multivariate split entropic coordinates}
\usage{
hasSplitSmetCoords(df)
}
\description{
Functions to detect SMET coordinates
a function to detect if the source multivariate split entropies are present: this enables 
working out the multivariate split entropic coordinates
}
