% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentropies.R
\name{sentropies.confusionMatrix}
\alias{sentropies.confusionMatrix}
\title{Entropy decomposition of a confusion table from caret}
\usage{
\method{sentropies}{confusionMatrix}(ct, ...)
}
\description{
Given a contingency matrix, provide one row of entropy coordinates. 
NOTE: the reference variable has to index the ROWS of the table, while the predicted
variable indexes the columns, unlike, e.g. \code{\link[caret]{confusionMatrix}}
}
