% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentropies.R
\name{sentropies.data.frame}
\alias{sentropies.data.frame}
\title{Multivariate source entropy decomposition of a data frame}
\usage{
\method{sentropies}{data.frame}(df, type = "total", ...)
}
\arguments{
\item{type}{The type of analysis being requested. Unles type=="dual" it will perform 
a normal analysis providing total + dual total correlation. If "dual" is used, then
only the dual total correlation will be provided. In this case, no individual split equations 
will be output, only a single aggregate, and the names of the columns are changed accordingly.}

\item{data}{The data being provided to the function.}
}
\value{
Another dataframe with the main entropy coordinates of every variable
  in the original, which are now the rows of the returned data.frame. If the columns have no
  names, artificial ones are returned based in pre prefix "x" and their column number. 
  The entropies are always in bits.
}
\description{
Multivariate source entropy decomposition of a data frame
}
