% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{simpleEntropies}
\alias{simpleEntropies}
\title{A characterization of the variables in a data.frame needed to define the perplexities.}
\format{\preformatted{ chr [1:5] "Ux" "Uy" "Hx" "Hy" "Hxy"
}}
\usage{
simpleEntropies
}
\description{
A characterization of the variables in a data.frame needed to define the perplexities.
}
\keyword{datasets}

